// libs/admin-shared/src/model/sysMenuTreeResponse.ts
/**
 * Generated by orval v6.11.1 🍺
 * Do not edit manually.
 * 工业资源boss管理端【API接口文档】
 * 工业资源boss管理端
 */
import type { SysMenuTreeResponsePermissionList } from './sysMenuTreeResponsePermissionList';

/**
 * 系统菜单表
 */
export interface SysMenuTreeResponse {
  /** 菜单英文别名(用于生成编号前缀) */
  alias?: string;
  /** 子孙节点 */
  children?: SysMenuTreeResponse[];
  /** 菜单编号(编号规则生成) */
  code?: string;
  /** 菜单图标 */
  icon?: string;
  /** 菜单ID */
  id?: string;
  /** 属性是否可维护（0否，1是） */
  isAttributeSet?: boolean;
  /** 菜单名称 */
  name?: string;
  /** 父级菜单编号 */
  parentCode?: string;
  /** 父级菜单id */
  parentId?: string;
  /** 请求地址 */
  path?: string;
  /** 地址类型{route:路由地址,url:外部url} */
  pathType?: string;
  /** 按钮权限 */
  permissionList?: SysMenuTreeResponsePermissionList;
  /** 显示位置 */
  showPosition?: string;
  /** sort */
  sort?: number;
  /** 所属系统 */
  system?: string;
  /** 目标{addTabPage:页签栏,_blank:新开窗口} */
  target?: string;
  /** 菜单类型{menu:菜单,button:按钮} */
  type?: string;
}